Servers: Add create() method for Physics2D query parameters.
There already existed a create() method for PhysicsRayQueryParameter2D,
but none of the other parameters had it.

Adding a create() method to every parameter provides consistency between
them, and enables functionality across the board that was previously only available on Physics2DRayQueryParameter.

Resolves godotengine/godot-proposals#8874

Servers: Add create() method for Physics3D query parameters.

Moved the PhysicsRayQueryParameter3D.create() method higher up in the
script to inhabit the same section as the rest of the methods for that
class.

There already existed a create() method for PhysicsRayQueryParameter3D,
but none of the other parameters had it.

Adding a create() method to every parameter provides consistency between
them, and enables functionality across the board that was previously
only available on Physics3DRayQueryParameter.